quantum_ml_time_series_project/src/quantum_feature_extraction.py

import numpy as np
from sklearn.decomposition import PCA
from sklearn.feature_selection import mutual_info_regression

class QuantumFeatureExtraction:
    def __init__(self, num_qubits):
        self.num_qubits = num_qubits

    def extract_principal_components(self, data, n_components=2):
        """Extracts the principal components using PCA for dimensionality reduction."""
        pca = PCA(n_components=n_components)
        return pca.fit_transform(data)

    def quantum_mutual_information(self, X, y):
        """Calculates mutual information between features and target using quantum-based methods."""
        return mutual_info_regression(X, y)

    def apply_quantum_entanglement(self, data):
        """Transforms features using quantum entanglement operations."""
        entangled_features = []
        for i in range(len(data)):
            entangled_row = [self._entangle(x) for x in data[i]]
            entangled_features.append(entangled_row)
        return np.array(entangled_features)

    def _entangle(self, x):
        """Simulates an entanglement transformation on a feature."""
        return np.sin(x) ** 2 + np.cos(x) ** 2

    def extract_temporal_quantum_features(self, time_series_data, look_back=3):
        """Extracts temporal quantum-specific features from time series data."""
        features = []
        for i in range(len(time_series_data) - look_back):
            segment = time_series_data[i:i + look_back]
            quantum_feature = np.mean(segment)  # Placeholder for more complex quantum extraction logic
            features.append(quantum_feature)
        return np.array(features)

    def save_features(self, features, file_path):
        """Saves extracted features to a file."""
        np.save(file_path, features)

    def load_features(self, file_path):
        """Loads saved features from a file."""
        return np.load(file_path)
